refactor: extract listMessages query logic into messageService
Moves the ~180-line GET /messages handler body (threaded CTE, non-threaded
query, cached total selection) into services/messageService.js so the logic
can be unit-tested without an HTTP layer.

Adds messageService.test.js with tests covering the non-obvious branching:
no-account early return, unowned accountId fallback to unified inbox,
unreadOnly toggling between unread_count and total_count in both the
per-account and unified-inbox paths.

Route handler becomes a thin shim: parse params, call listMessages, trigger
prefetch side-effect, send response.
